main component changes
