+++ src/components/App.jsx
@@ -1,0 +1,6 @@
+function App() {
+
+  
+}
+
+export default App;
