+++ src/index.js
@@ -3,11 +3,14 @@ import ReactDOM from 'react-dom/client';
 import './styles/index.scss';
 import App from './components/App';
 import Header from './components/Header';
+import Body from './components/Body';
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Header />
     <App />
+    <Header />
+    <Body/>
   </React.StrictMode>
 );
