+++ src/styles/body.scss
@@ -36,7 +36,13 @@
     display: flex;
     align-items: center;
     justify-content: center;
+    gap: 20px;
     margin: 10%;
+    padding: 10%;
+    background-image: url('../images/bg-shorten-desktop.svg');
+    background-color: styles.$Very-Dark-Violet;
+    background-size: contain;
+    background-repeat: no-repeat;
 }
 
 .secondary-container {
